Add files via upload
Coloraise is a plugin which allows users to change the colors on a web page, and the size of text elements. This plugin can be used anywhere it is desirable to allow users to change colors on a page on a more granular level than the high contrast setting on Windows or the inverted colors setting on Mac OS X, but it is meant to extend accessibility to users with low vision or color blindness.  The web designer has complete control over what can and cannot change colors, through the use of a mini-framework of classes which are described in detail at aDuboisForge.com.